Another major commit to finally share transport protocols properly among different containers, and make external port support independent of any particular container. This change moves more functionality out of the RCC container into the base class used by all containers.  The major motivator was that HDL containers now support multiple protocols, and not doing this would have required a lot of redundant code. Now the code is properly shared so that different container types have an easier time using different transport protocols. We also added the g++ printf attributes to all functions, methods, exceptions, etc. that take printf variable arguments so that they are all now checked for consistency - a big win.
Makefile: another swipe at building ocpiocl in the right cases of shared libraries.
Makefile.ocpi.common: get rid of ancient incorrect dependency tracking of idl files
Makefile.ocpi.for-idl: ditto
adapt/os/ocpios/interfaces/include/OcpiOsDebug.h: enable g++ to check printf strings and args in ocpiDebug
adapt/os/ocpios/linux/src/OcpiOsEther.cxx: one of many files touched due to new printf arg warnings
-- no more files will be mentioned that simply fixed newly enabled warnings for ocpiDebug etc.
core/container/ctests/bin/src/GPPContainerTest_main.cxx: new packportdesc interface update
core/container/ctests/bin/src/RCCLbTest_main.cxx: ditto
core/container/ctests/bin/src/test_control_errors_main.cxx: some tests were obsolete and invalid
-- among other things, starting a started worker and stopping a stopped worker is ok now.
-- improve test output to name individual tests
core/container/ctests/bin/src/test_user_port_main.cxx: test using API not internal
core/container/ctests/workers/src/ProdWorker.c: nuke warning
core/container/hdl_container/Makefile.ocpi.imports: include ether datagram transports for testing
core/container/hdl_container/bin/src/ocpihdl_main.cxx: add ether datagram RDMA testing with HW
core/container/hdl_container/impl/include/HdlAccess.h: improve endpoint prep for container
core/container/hdl_container/impl/include/HdlContainer.h: ditto
core/container/hdl_container/impl/src/EtherDriver.cxx: ditto
core/container/hdl_container/impl/src/PciDriver.cxx: ditto
core/container/hdl_container/impl/src/HdlContainer.cxx: ditto
-- and remove code to share new external port code
core/container/hdl_container/impl/include/HdlOCCP.h: warning
core/container/hdl_container/impl/include/HdlOCDP.h: enable for C++ namespaces
core/container/interfaces/Makefile.ocpi.imports: depend on transport as well as transfer libs
core/container/interfaces/api/include/OcpiContainerApi.h: more consistency of "params" vs "props"
core/container/interfaces/api/include/OcpiContainerInterface.h: have base class be a time emitter
-- move port descriptor packing and unpacking to the port file
-- move transport session management into the base class
core/container/interfaces/api/src/OcpiContainerInterface.cxx: ditto
core/container/interfaces/api/include/OcpiContainerManager.h: transport global management is here now
-- it was in the rcc driver, now it is shared.
core/container/interfaces/api/src/OcpiContainerManager.cxx: ditto
core/container/interfaces/api/include/OcpiContainerPort.h: big changes to absorb external ports
-- put the mutex in the base class
-- adjust connection interface to increase shared code and external port code being generic
-- put port packing code here
core/container/interfaces/api/src/OcpiContainerPort.cxx: ditto, plus
-- don't mess with packed port descriptors when the containers are local
core/container/interfaces/api/include/OcpiWorker.h: allow ports to retrive worker names
core/container/interfaces/api/src/OcpiWorker.cxx: allow redundant control operations, and suppress them
core/container/ocl_container/impl/src/OcpiOclContainer.cxx: update for new interfaces and external ports
core/container/ocl_container/impl/src/OcpiOclPlatformManager.cxx: improve exception messages
core/container/rcc_container/include/OcpiRccContainer.h:  many moves into base classes
-- time::emit is in base class for all containers
-- transport object management (transport session) moved to base classes
core/container/rcc_container/src/OcpiRccContainer.cxx: ditto
core/container/rcc_container/include/OcpiRccDriver.h: transport global management moved to container manager
core/container/rcc_container/src/OcpiRccDriver.cxx: ditto
core/container/rcc_container/include/OcpiRccPort.h: big changes for moving functionality into base class
core/container/rcc_container/src/OcpiRccPort.cxx: ditto
core/container/rcc_container/src/OcpiRccMessagePort.cxx: tweaks for new connectivity interfaces
core/container/rcc_container/src/OcpiRccWorker.cxx: a better job of checking optional ports
-- when "stop" fails due to the worker returning an error, the worker is not stopped.
======= A number of files changes to ensure that endpoint ids (a.k.a. mailbox numbers) are uniformly uint16_t
core/dataplane/rdma_driver_interface/interfaces/include/DtHandshakeControl.h: ditto
core/dataplane/rdma_driver_interface/interfaces/include/DtSharedMemoryInterface.h: ditto
core/dataplane/rdma_driver_interface/interfaces/src/DtSharedMemoryInternal.cxx: ditto
core/dataplane/rdma_driver_interface/interfaces/include/DtTransferInterface.h: ditto
core/dataplane/rdma_driver_interface/interfaces/include/DtTransferInternal.h: ditto
core/dataplane/rdma_drivers/ofed/src/DtOfedXfer.cxx: ditto
core/dataplane/rdma_drivers/pci/include/DtPCIPioXfer.h: ditto
core/dataplane/rdma_drivers/pci/src/DtPCIPioXfer.cxx: ditto
core/dataplane/rdma_drivers/pio/include/DtPioXfer.h: ditto
core/dataplane/rdma_drivers/pio/src/DtPioXfer.cxx: ditto
core/dataplane/rdma_drivers/sockets/include/DtSocketsXfer.h: ditto
core/dataplane/rdma_drivers/sockets/src/DtSocketsXfer.cxx: ditto
core/dataplane/rdma_drivers/udp/include/DtUDPSocketXfer.h: ditto
core/dataplane/rdma_drivers/udp/src/DtUDPSocketXfer.cxx: ditto
core/dataplane/rdma_driver_interface/interfaces/include/DtTransferInternal.h: ditto
-- also add allocateProxyEndPoint function to allow hardware containers to generate endpoints properly
core/dataplane/rdma_driver_interface/interfaces/src/DtTransferInternal.cxx: ditto
core/dataplane/rdma_driver_interface/interfaces/include/OcpiRDTInterface.h: start nuking "RDT"
core/dataplane/rdma_drivers/datagram/include/DtDataGramXfer.h: 16 bit endpoints, correct signness
core/dataplane/rdma_drivers/datagram/src/DtDataGramEtherDriver.cxx: ditto
core/dataplane/rdma_drivers/datagram/src/DtDataGramUDPDriver.cxx: ditto
core/dataplane/rdma_drivers/datagram/src/DtDataGramXfer.cxx: ditto, plus a real boundary bug fix
core/dataplane/transport/impl/include/OcpiPort.h: adjust finalize interface to return correct feedback
-- i.e. relieve the caller of knowing which type of descriptor should be used afterward
core/dataplane/transport/impl/src/OcpiPort.cxx: ditto
-- also try to initialize the circuit eariler when possible
core/dataplane/transport/impl/include/OcpiTransport.h: provide an iterator typedef, use list rather than vector
core/dataplane/transport/impl/src/OcpiTransport.cxx: use list rather than vector to avoid holes, 16 bit mboxes
core/dataplane/transport/impl/src/OcpiCircuit.cxx: better debug messages for port closure
core/library/src/OcpiLibraryAssembly.cxx: printf warnings fixed due to new attributes
core/local/util/misc/include/OcpiUtilException.h: add g++ attributes for printf checking
core/local/util/misc/include/OcpiUtilMisc.h: ditto
core/local/util/misc/include/OcpiUtilSelfMutex.h: add SelfRefMutex when you want a reference
-- this makes this case the same API as the SelfMutex (inherited mutex) case.
core/local/util/misc/src/OcpiUtilException.cxx: make debug logging of exceptions clearer
core/local/util/parentChild/include/OcpiParentChild.h: make parent retrieval a const method
test/hdl/scripts/loadBitStreamOnPlatform: strip "PCI:" prefix that is now used
test/hdl/scripts/probeXilinxUsb: use listusbcables function rather than trying a bunch of names

+++ core/dataplane/rdma_driver_interface/interfaces/include/DtTransferInterface.h
@@ -285,7 +285,7 @@ namespace DataTransfer {
     inline EndPoint* getEndPoint(const std::string &s, bool local=false) {
       return getEndPoint(s.c_str(), local);
     }
-    virtual EndPoint* addCompatibleEndPoint( uint32_t mbox, uint32_t maxMb);
+    virtual EndPoint* addCompatibleEndPoint( uint16_t mbox, uint16_t maxMb);
     virtual EndPoint* addEndPoint(const char *endpoint, bool local);
     // Avoid the mailbox, and match the mailbox count, if not -1
     virtual EndPoint* createEndPoint(std::string& endpoint, bool local=false) = 0;
@@ -297,7 +297,7 @@ namespace DataTransfer {
      * it has been passed to finalizeEndpoint().
      ***************************************/
     virtual std::string allocateEndpoint(const OCPI::Util::PValue*,
-					 unsigned mailBox, unsigned maxMailBoxes) = 0;
+					 uint16_t mailBox, uint16_t maxMailBoxes) = 0;
 
     // The endpoint is telling its factory that it is being deleted.
     void removeEndPoint(EndPoint &ep);
@@ -328,7 +328,7 @@ namespace DataTransfer {
     static ezxml_t getNode( ezxml_t tn, const char* name );
     // The range of mailboxes is common across all transports.
     // This will allow us to share memory between protocol someday.
-    uint32_t getMaxMailBox(), getNextMailBox();
+    uint16_t getMaxMailBox(), getNextMailBox();
   private:
     // This vector is indexed by the mailbox number of the endpoint
     // and only contains local endpoints
